Move some models and helpers into search common
These are pieces we'll want to reuse in the snapshot generator.  This
just moves them into the common lib so they're available in that app.
