+++ common/search/src/main/scala/weco/api/search/models/index/IndexedWork.scala
@@ -1,15 +1,16 @@
 package weco.api.search.models.index
 
 import io.circe.Json
-import weco.api.search.json.CatalogueJsonUtil
-import weco.api.search.models.request.WorksIncludes
+import io.circe.syntax._
+import weco.catalogue.display_model.work.DisplayWork
+import weco.catalogue.display_model.Implicits._
 import weco.catalogue.internal_model.work.{Work, WorkState}
 
 sealed trait IndexedWork
 
 case class Identified(canonicalId: String)
 
-object IndexedWork extends CatalogueJsonUtil {
+object IndexedWork {
   case class Visible(display: Json) extends IndexedWork
 
   // Note: this layer of indirection is because of the current index structure,
@@ -20,18 +21,10 @@ object IndexedWork extends CatalogueJsonUtil {
   case class Invisible() extends IndexedWork
   case class Deleted() extends IndexedWork
 
-  // TODO: These are temporary methods used while we're moving the API to
-  // use the "display" field in the index; once that work is done we can
-  // delete this.
-  object Visible {
-    def apply(work: Work.Visible[WorkState.Indexed]): IndexedWork.Visible =
-      IndexedWork.Visible(work.asJson(WorksIncludes.all))
-  }
-
   def apply(work: Work[WorkState.Indexed]): IndexedWork =
     work match {
       case w: Work.Visible[WorkState.Indexed] =>
-        IndexedWork.Visible(w)
+        IndexedWork.Visible(DisplayWork(w).asJson)
 
       case w: Work.Redirected[WorkState.Indexed] =>
         IndexedWork.Redirected(
