+++ search/src/test/scala/weco/api/search/ApiSearchTemplatesTest.scala
@@ -3,6 +3,7 @@ package weco.api.search
 import akka.http.scaladsl.model.ContentTypes
 import io.circe.Json
 import org.scalatest.matchers.should.Matchers
+import weco.api.search.fixtures.JsonHelpers
 import weco.api.search.works.ApiWorksTestBase
 
 class ApiSearchTemplatesTest
